Library

Documentation for AstroForceModels.jl.

AstroForceModels.AbstractSatelliteDragModel — Type

Abstract satellite drag model to help compute drag forces.

AstroForceModels.AbstractSatelliteSRPModel — Type

Abstract satellite drag model to help compute drag forces.

AstroForceModels.CannonballFixedDrag — Type

Cannonball Fixed Drag Model struct Contains information to compute the ballistic coefficient of a cannonball drag model with a fixed ballistic coefficient.

Fields

radius::Number: The radius of the spacecraft.

mass::Number: The mass of the spacecraft.

drag_coeff::Number: The drag coefficient of the spacecraft.

ballistic_coeff::Number: The fixed ballistic coefficient to use.

AstroForceModels.CannonballFixedDrag — Method

CannonballFixedDrag(radius::Number, mass::Number, drag_coeff::Number)

Constructor for a fixed cannonball ballistic coefficient drag model.

The ballistic coefficient is computed with the following equation:

            BC = CD * area/mass

where area is the 2D projection of a sphere

            area = π * r^2

Arguments

radius::Number: The radius of the spacecraft.

mass::Number: The mass of the spacecraft.

drag_coeff::Number: The drag coefficient of the spacecraft.

Returns

drag_model::CannonballFixedDrag: A fixed ballistic coefficient drag model.

AstroForceModels.CannonballFixedDrag — Method

CannonballFixedDrag(ballistic_coeff::Number)

Constructor for a fixed ballistic coefficient drag model.

Arguments

ballistic_coeff::Number: The fixed ballistic coefficient to use.

Returns

drag_model::CannonballFixedDrag: A fixed ballistic coefficient drag model.

AstroForceModels.CannonballFixedSRP — Type

Cannonball Fixed SRP Model struct Contains information to compute the reflectivity coefficient of a cannonball drag model with a fixed reflectivity coefficient.

Fields

radius::Number: The radius of the spacecraft.

mass::Number: The mass of the spacecraft.

reflectivity_coeff::Number: The reflectivity coefficient of the spacecraft.

reflectivity_ballistic_coeff::Number: The fixed ballistic coefficient to use.

AstroForceModels.CannonballFixedSRP — Method

CannonballFixedSRP(radius::Number, mass::Number, drag_coeff::Number)

Constructor for a fixed cannonball ballistic coefficient SRP model.

The ballistic coefficient is computed with the following equation:

            RC = CD * area/mass

where area is the 2D projection of a sphere

            area = π * r^2

Arguments

radius::Number: The radius of the spacecraft.

mass::Number: The mass of the spacecraft.

reflectivity_coeff::Number: The reflectivity coefficient of the spacecraft.

Returns

srp_model::CannonballFixedSRP`: A fixed ballistic coefficient SRP model.

AstroForceModels.CannonballFixedSRP — Method

CannonballFixedSRP(ballistic_coeff::Number)

Constructor for a fixed ballistic coefficient SRP model.

Arguments

ballistic_coeff::Number: The fixed ballistic coefficient to use.

Returns

srp_model::CannonballFixedDrag: A fixed ballistic coefficient SRP model.

AstroForceModels.CelestialBody — Method

Creates a CelestialBody Object from a name Available: {Sun, Mercury, Venus, Earth, Mars, Jupiter, Saturn, Uranus, Neptune, Pluto, }

Arguments:

name::String: Name of the Planet Object (Casing Doesn't Matter)

(Optional) T::DataType`: Datatype of PlanetBody Fields

Returns: -CelestialBody: PlanetBody Object with Fields: - name::String: Name of Object - central_body::String: Name of Central Body Being Orbited - jpl_code::Int: NAIF ID Code - ephemeris::Function: Ephemeris Object

- μ::T: Graviational Parameter [km/s]
- Req::T: Equatorial Radius [km]

AstroForceModels.DragAstroModel — Type

Drag Astro Model struct Contains information to compute the acceleration of a drag force on a spacecraft.

Fields

satellite_drag_model::AbstractSatelliteDragModel: The satellite drag model for computing the ballistic coefficient.

atmosphere_model::Symbol: The atmospheric model for computing the density.

eop_data::EopIau1980: Earth orientation parameters to help compute the density with the atmospheric model.

AstroForceModels.SRPAstroModel — Type

SRP Astro Model struct Contains information to compute the acceleration of a SRP a spacecraft.

Fields

satellite_srp_model::AbstractSatelliteDragModel: The satellite srp model for computing the ballistic coefficient.

sun_data::ThirdBodyModel: The data to compute the Sun's position.

AstroForceModels.StateDragModel — Type

State Drag Model struct Empty struct used for when the simulation state includes the ballistic coefficient.

AstroForceModels.StateSRPModel — Type

State SRP Model struct Empty struct used for when the simulation state includes the reflectivity ballistic coefficient.

AstroForceModels.CustomCelestialBody — Method

Creates Custom CelestialBody Object from a name

Arguments: -'name::String': Name of the Planet Object (Casing Doesn't Matter) -'ephemeris::Function': Ephemeris function -(Optional) 'centralbody::String': Name of Central Body Being Orbited -(Optional) 'jplcode::Int64': NAIF ID Code -(Optional) 'μ::Number': Graviational Parameter [km/s] -(Optional) 'Req::Number': Equatorial Radius [km]

Returns: -'CelestialBody': PlanetBody Object with Fields: name::String: Name of Object centralbody::String: Name of Central Body Being Orbited jplcode::Int64: NAIF ID Code ephemeris::Function: Ephemeris function μ::Number: Graviational Parameter [km/s] Req::Number: Equatorial Radius [km]

AstroForceModels.acceleration — Method

' acceleration(drag_model::DragAstroModel, u::AbstractArray, p::ComponentVector, t::Number)

Computes the drag acceleration acting on a spacecraft given a drag model and current state and parameters of an object.

Arguments

u::AbstractArray: Current State of the simulation.

p::ComponentVector: Current parameters of the simulation.

t::Number: Current time of the simulation.

drag_model::DragAstroModel: Drag model struct containing the relevant information to compute the acceleration.

Returns

acceleration: SVector{3}: The 3-dimensional drag acceleration acting on the spacecraft.

AstroForceModels.acceleration — Method

' acceleration(drag_model::DragAstroModel, u::AbstractArray, p::ComponentVector, t::Number)

Computes the drag acceleration acting on a spacecraft given a drag model and current state and parameters of an object.

Arguments

u::AbstractArray: Current State of the simulation.

p::ComponentVector: Current parameters of the simulation.

t::Number: Current time of the simulation.

srp_model::DragAstroModel: Drag model struct containing the relevant information to compute the acceleration.

Returns

acceleration: SVector{3}: The 3-dimensional srp acceleration acting on the spacecraft.

AstroForceModels.ballistic_coefficient — Method

ballistic_coefficient( u::AbstractArray, p::ComponentVector, t::Number, model::CannonballFixedDrag)

Returns the ballistic coefficient for a drag model given the model and current state of the simulation.

Arguments

- `u::AbstractArray`: The current state of the simulation.
- `p::ComponentVector`: The parameters of the simulation.
- `t::Number`: The current time of the simulation.
- `model::CannonballFixedDrag`: Drag model for the spacecraft.

Returns

-`ballistic_coeff::Number`: The current ballistic coefficient of the spacecraft.

AstroForceModels.ballistic_coefficient — Method

ballistic_coefficient( u::AbstractArray, p::ComponentVector, t::Number, model::StateDragModel)

Returns the ballistic coefficient for a drag model given the model and current state of the simulation.

Arguments

- `u::AbstractArray`: The current state of the simulation.
- `p::ComponentVector`: The parameters of the simulation.
- `t::Number`: The current time of the simulation.
- `model::StateDragModel`: The drag model for the spacecraft.

Returns

-`ballistic_coeff::Number`: The current ballistic coefficient of the spacecraft.

AstroForceModels.compute_density — Function

compute_density(JD::Number, u::AbstractArray, eop_data::EopIau1980, Val(AtmosphereType::Symbol))

Computes the atmospheric density at a point given the date, position, eop_data, and atmosphere type

Arguments

JD::Number: The current time of the simulation in Julian days.

u::AbstractArray: The current state of the simulation.

eop_data::EopIau1980: The earth orientation parameters.

AtmosphereType::Symbol: The type of atmospheric model used to compute the density. Available options are Jacchia-Bowman 2008 (:JB2008), Jacchia-Roberts 1971 (:JR1971), NRL MSIS 2000 (:MSIS2000), Exponential (:ExpAtmo), and None (:None)

Returns

rho::Number: The density of the atmosphere at the provided time and point [kg/m^3].

AstroForceModels.drag_accel — Method

drag_accel(u::AbstractArray, rho::Number, BC::Number, ω_vec::AbstractArray, t::Number, [DragModel]) -> SVector{3}{Number}

Compute the Acceleration Atmospheric Drag

The atmosphere is treated as a solid revolving with the Earth and the apparent velocity of the satellite is computed using the transport theorem

            𝐯_app = 𝐯 - 𝛚 x 𝐫

The acceleration from drag is then computed with a cannonball model as

            𝐚 = 1/2 * ρ * BC * |𝐯_app|₂^2 * v̂

Note

Currently only fixed cannonball state based ballistic coefficients are supported, custom models can be created for higher fidelity.

Arguments

u::AbstractArray: The current state of the spacecraft in the central body's inertial frame.

rho::Number: Atmospheric density at (t, u) [kg/m^3].

BC::Number: The ballistic coefficient of the satellite – (area/mass) * drag coefficient [kg/m^2].

ω_vec::AbstractArray: The angular velocity vector of Earth. Typically approximated as [0.0; 0.0; ω_Earth]

t::Number: Current time of the simulation.

Returns

SVector{3}{Number}: Inertial acceleration from drag

AstroForceModels.reflectivity_ballistic_coefficient — Method

reflectivityballisticcoefficient( u::AbstractArray, p::ComponentVector, t::Number, model::CannonballFixedSRP)

Returns the ballistic coefficient for a SRP model given the model and current state of the simulation.

Arguments

- `u::AbstractArray`: The current state of the simulation.
- `p::ComponentVector`: The parameters of the simulation.
- `t::Number`: The current time of the simulation.
- `model::CannonballFixedSRP`: SRP model for the spacecraft.

Returns

-`ballistic_coeff::Number`: The current ballistic coefficient of the spacecraft.

AstroForceModels.reflectivity_ballistic_coefficient — Method

ballistic_coefficient( u::AbstractArray, p::ComponentVector, t::Number, model::StateSRPModel)

Returns the ballistic coefficient for a SRP model given the model and current state of the simulation.

Arguments

- `u::AbstractArray`: The current state of the simulation.
- `p::ComponentVector`: The parameters of the simulation.
- `t::Number`: The current time of the simulation.
- `model::StateSRPModel`: The SRP model for the spacecraft.

Returns

-`ballistic_coeff::Number`: The current ballistic coefficient of the spacecraft.

AstroForceModels.relativity_accel — Method

AstroForceModels.shadow_model — Method

shadow_model(sat_pos::AbstractArray, sun_pos::AbstractArray, R_Sun::Number, R_Earth::Number, t::Number, ShadowModel::Symbol)

Computes the Lighting Factor of the Sun occur from the Umbra and Prenumbra of Earth's Shadow

Arguments

sat_pos::AbstractArray: The current satellite position.

sun_pos::AbstractArray: The current Sun position.

R_Sun::Number: The radius of the Sun.

R_Earth::Number: The radius of the Earth.

ShadowModel::Symbol: The Earth shadow model to use. Current Options – :Cylindrical, :Conical, :Conical_Simplified, :None

Returns

SVector{3}{Number}: Inertial acceleration from the 3rd body

AstroForceModels.srp_accel — Method

srp_accel(u::AbstractArray, sun_pos::AbstractArray, R_Sun::Number, R_Earth::Number, Ψ::Number, RC::Number, t::Number, [SRPModel::Symbol]; ShadowModel::Symbol)

Compute the Acceleration from Solar Radiaiton Pressure

Radiation from the Sun reflects off the satellite's surface and transfers momentum perturbing the satellite's trajectory. This force can be computed using the a Cannonball model with the following equation

            𝐚 = F * RC * Ψ * (AU/(R_sc_Sun))^2 * R̂_sc_Sun

Note

Currently only Cannonball SRP is supported, to use a higher fidelity drag either use a state varying function or compute the ballistic coefficient further upstream

Arguments

u::AbstractArray: The current state of the spacecraft in the central body's inertial frame.

sun_pos::AbstractArray: The current position of the Sun.

R_Sun::Number: The radius of the Sun.

R_Earth::Number: The radius of the Earth.

Ψ::Number: Solar Constant at 1 Astronomical Unit.

RC::Number: The solar ballistic coefficient of the satellite – (Area/mass) * Reflectivity Coefficient [kg/m^2].

t::Number: The current time of the Simulation

Optional Arguments

SRPModel::Symbol: SRP Model to use. Current Options – :Cannonball, :None

ShadowModel::Symbol: SRP Earth Shadow Model to use. Current Options – :Conical, :Conical_Simplified, :Cylinderical

Returns

SVector{3}{Number}: Inertial acceleration from the 3rd body

AstroForceModels.third_body_accel — Method

third_body_accel(u::AbstractArray, μ_body::Number, body_pos::AbstractArray, h::Number) -> SVector{3}{Number}

Compute the Acceleration from a 3rd Body Represented as a Point Mass

Since the central body is also being acted upon by the third body, the acceleration of body 𝐁 acting on spacecraft 𝐀 in the orbiting body's 𝐂 is part of the force not acting on the central body

            a = ∇UB(rA) - ∇UB(rC)

Arguments

u::AbstractArray: The current state of the spacecraft in the central body's inertial frame.

μ_body: Gravitation Parameter of the 3rd body.

body_pos::AbstractArray: The current position of the 3rd body in the central body's inertial frame.

Returns

SVector{3}{Number}: Inertial acceleration from the 3rd body